Moved a few functions from appiumHelper
