+++ test/smoke/src/helpers/appiumHelper.ts
@@ -82,39 +82,6 @@ export class AppiumHelper {
         return wdio.remote(attachArgs);
     }
 
-    // Check if appPackage is installed on Android device for waitTime ms
-    public static async checkIfAppIsInstalled(appPackage: string, waitTime: number, waitInitTime?: number) {
-        let awaitRetries: number = waitTime / 1000;
-        let retry = 1;
-        await new Promise((resolve, reject) => {
-            let check = setInterval(async () => {
-                if (retry % 5 === 0) {
-                    console.log(`*** Check if app is being installed with command 'adb shell pm list packages ${appPackage}' for ${retry} time`);
-                }
-                let result;
-                try {
-                    result = cp.execSync(`adb shell pm list packages ${appPackage}`).toString().trim();
-                } catch (e) {
-                    clearInterval(check);
-                    reject(`Error occured while check app is installed:\n ${e}`);
-                }
-                if (result) {
-                    clearInterval(check);
-                    const initTimeout = waitInitTime || 10000;
-                    console.log(`*** Installed ${appPackage} app found, await ${initTimeout}ms for initializing...`);
-                    await sleep(initTimeout);
-                    resolve();
-                } else {
-                    retry++;
-                    if (retry >= awaitRetries) {
-                        clearInterval(check);
-                        reject(`${appPackage} not found after ${waitTime}ms`);
-                    }
-                }
-            }, 1000);
-        });
-    }
-
     public static async openExpoApplicationAndroid(client: WebdriverIO.Client<WebdriverIO.RawResult<null>> & WebdriverIO.RawResult<null>, expoURL: string) {
         if (process.platform === "darwin") {
             // Longer way to open Expo app, but
